Create PickUp and get data from frontend and store in database
