+++ src/main/java/com/example/courier304project/controller/ParcelController.java
@@ -1,28 +1,25 @@
 package com.example.courier304project.controller;
 
 import com.example.courier304project.dto.ParcelDto;
-import com.example.courier304project.entity.Customer;
+import com.example.courier304project.entity.Parcel;
+import com.example.courier304project.service.ParcelService;
 import com.example.courier304project.service.ParcelService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
+@CrossOrigin(origins = "http://localhost:3000")
 @RestController
-@RequestMapping("/api/v1")
+@RequestMapping("/user")
 public class ParcelController {
 
     @Autowired
     private ParcelService parcelService;
 
-    @PostMapping("addParcel")
-    public void addParcel(@RequestBody ParcelDto parcelDto){
-        parcelService.addParcel(parcelDto);
-    }
 
-    @GetMapping("/parcel")
-    public String getAllParcel(){
-        return "parcel controller";
-    }
 
+    @PostMapping("/addparcel")
+    public Parcel addParcel (@RequestBody ParcelDto parcelDto){
+       return parcelService.addParcel(parcelDto);
+
+   }
 }
