+++ src/main/java/com/example/courier304project/entity/Parcel.java
@@ -1,142 +1,228 @@
 package com.example.courier304project.entity;
 
-import jakarta.persistence.*;
-
-import java.sql.Date;
-import java.sql.Time;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 
 @Entity
 public class Parcel {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int parcelId;
-    private Float cost;
-    private int paymentType;
-    private float length;
-    private float width;
-    private float height;
-    private Date date;
-    private Time time;
-    private int scheduleNo;
-    private double distance;
-    private int status; // 1= one day delevery, 2= 3 days,  3= within week , 4 = other
+    Long ParcelID;
 
-    //@OneToMany(targetEntity = Parcel.class, cascade = CascadeType.ALL)
-    //@JoinColumn(name="fk",referencedColumnName = "customerPhone")
+    String senderPhoneNo;
+    String senderName;
+    String senderEmail;
+    String senderAddress;
+    String senderCity;
+    String senderDistrict;
 
+    String receiverPhoneNo;
+    String receiverName;
+    String receiverAddress;
+    String receiverDistrict;
+    String receiverCity;
 
-    @ManyToOne
-    //@JoinColumn(name="fk",referencedColumnName = "customerPhone")
-    private Customer customer;
+    //"12:12 AM" 'Date("11/24/2021")'
+    String timeFrom;
+    String timeTo;
+    String pickupDate;
+    String vehicleType;
+    String postMethod;
+    String paymentType;
+    String specialNote; //spetial
 
-    public Parcel(){}
+    int cost;
 
-    public Customer getCustomer() {
-        return customer;
+    public Parcel() {
     }
 
-    public void setCustomer(Customer customer) {
-        this.customer = customer;
-    }
-
-    //constructor only cost,paymentType,dimension,status
-    public Parcel(Float cost, int paymentType, float length, float width, float height, int status,Customer customer) {
-        this.cost = cost;
+    public Parcel(String senderPhoneNo, String senderName, String senderEmail,
+                  String senderAddress, String senderCity, String senderDistrict,
+                  String receiverPhoneNo, String receiverName, String receiverAddress,
+                  String receiverDistrict, String receiverCity, String timeFrom, String timeTo,
+                  String pickupDate, String vehicleType, String postMethod, String paymentType,
+                  String specialNote,int cost) {
+        this.senderPhoneNo = senderPhoneNo;
+        this.senderName = senderName;
+        this.senderEmail = senderEmail;
+        this.senderAddress = senderAddress;
+        this.senderCity = senderCity;
+        this.senderDistrict = senderDistrict;
+        this.receiverPhoneNo = receiverPhoneNo;
+        this.receiverName = receiverName;
+        this.receiverAddress = receiverAddress;
+        this.receiverDistrict = receiverDistrict;
+        this.receiverCity = receiverCity;
+        this.timeFrom = timeFrom;
+        this.timeTo = timeTo;
+        this.pickupDate = pickupDate;
+        this.vehicleType = vehicleType;
+        this.postMethod = postMethod;
         this.paymentType = paymentType;
-        this.length = length;
-        this.width = width;
-        this.height = height;
-        this.status = status;
-        this.customer= customer;
+        this.specialNote = specialNote;
+        this.cost=cost;
     }
 
-
-    public int getParcelId() {
-        return parcelId;
+    public Long getParcelID() {
+        return ParcelID;
     }
 
-    public void setParcelId(int parcelId) {
-        this.parcelId = parcelId;
+    public void setParcelID(Long parcelID) {
+        ParcelID = parcelID;
     }
 
-    public Float getCost() {
+    public int getCost() {
         return cost;
     }
 
-    public void setCost(Float cost) {
+    public void setCost(int cost) {
         this.cost = cost;
     }
 
-    public int getPaymentType() {
+    public String getSenderPhoneNo() {
+        return senderPhoneNo;
+    }
+
+    public void setSenderPhoneNo(String senderPhoneNo) {
+        this.senderPhoneNo = senderPhoneNo;
+    }
+
+    public String getSenderName() {
+        return senderName;
+    }
+
+    public void setSenderName(String senderName) {
+        this.senderName = senderName;
+    }
+
+    public String getSenderEmail() {
+        return senderEmail;
+    }
+
+    public void setSenderEmail(String senderEmail) {
+        this.senderEmail = senderEmail;
+    }
+
+    public String getSenderAddress() {
+        return senderAddress;
+    }
+
+    public void setSenderAddress(String senderAddress) {
+        this.senderAddress = senderAddress;
+    }
+
+    public String getSenderCity() {
+        return senderCity;
+    }
+
+    public void setSenderCity(String senderCity) {
+        this.senderCity = senderCity;
+    }
+
+    public String getSenderDistrict() {
+        return senderDistrict;
+    }
+
+    public void setSenderDistrict(String senderDistrict) {
+        this.senderDistrict = senderDistrict;
+    }
+
+    public String getReceiverPhoneNo() {
+        return receiverPhoneNo;
+    }
+
+    public void setReceiverPhoneNo(String receiverPhoneNo) {
+        this.receiverPhoneNo = receiverPhoneNo;
+    }
+
+    public String getReceiverName() {
+        return receiverName;
+    }
+
+    public void setReceiverName(String receiverName) {
+        this.receiverName = receiverName;
+    }
+
+    public String getReceiverAddress() {
+        return receiverAddress;
+    }
+
+    public void setReceiverAddress(String receiverAddress) {
+        this.receiverAddress = receiverAddress;
+    }
+
+    public String getReceiverDistrict() {
+        return receiverDistrict;
+    }
+
+    public void setReceiverDistrict(String receiverDistrict) {
+        this.receiverDistrict = receiverDistrict;
+    }
+
+    public String getReceiverCity() {
+        return receiverCity;
+    }
+
+    public void setReceiverCity(String receiverCity) {
+        this.receiverCity = receiverCity;
+    }
+
+    public String getTimeFrom() {
+        return timeFrom;
+    }
+
+    public void setTimeFrom(String timeFrom) {
+        this.timeFrom = timeFrom;
+    }
+
+    public String getTimeTo() {
+        return timeTo;
+    }
+
+    public void setTimeTo(String timeTo) {
+        this.timeTo = timeTo;
+    }
+
+    public String getPickupDate() {
+        return pickupDate;
+    }
+
+    public void setPickupDate(String pickupDate) {
+        this.pickupDate = pickupDate;
+    }
+
+    public String getVehicleType() {
+        return vehicleType;
+    }
+
+    public void setVehicleType(String vehicleType) {
+        this.vehicleType = vehicleType;
+    }
+
+    public String getPostMethod() {
+        return postMethod;
+    }
+
+    public void setPostMethod(String postMethod) {
+        this.postMethod = postMethod;
+    }
+
+    public String getPaymentType() {
         return paymentType;
     }
 
-    public void setPaymentType(int paymentType) {
+    public void setPaymentType(String paymentType) {
         this.paymentType = paymentType;
     }
 
-    public float getLength() {
-        return length;
+    public String getSpecialNote() {
+        return specialNote;
     }
 
-    public void setLength(float length) {
-        this.length = length;
-    }
-
-    public float getWidth() {
-        return width;
-    }
-
-    public void setWidth(float width) {
-        this.width = width;
-    }
-
-    public float getHeight() {
-        return height;
-    }
-
-    public void setHeight(float height) {
-        this.height = height;
-    }
-
-    public Date getDate() {
-        return date;
-    }
-
-    public void setDate(Date date) {
-        this.date = date;
-    }
-
-    public Time getTime() {
-        return time;
-    }
-
-    public void setTime(Time time) {
-        this.time = time;
-    }
-
-    public int getScheduleNo() {
-        return scheduleNo;
-    }
-
-    public void setScheduleNo(int scheduleNo) {
-        this.scheduleNo = scheduleNo;
-    }
-
-    public double getDistance() {
-        return distance;
-    }
-
-    public void setDistance(double distance) {
-        this.distance = distance;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public void setStatus(int status) {
-        this.status = status;
+    public void setSpecialNote(String specialNote) {
+        this.specialNote = specialNote;
     }
 }
