+++ src/main/java/com/example/courier304project/repository/ParcelRepository.java
@@ -1,17 +1,11 @@
 package com.example.courier304project.repository;
 
-
-import com.example.courier304project.entity.Address;
-import com.example.courier304project.entity.Customer;
 import com.example.courier304project.entity.Parcel;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 @Repository
-public interface ParcelRepository extends JpaRepository<Parcel,Integer> {
-
+public interface ParcelRepository extends JpaRepository<Parcel,Long> {
 
 
 }
