--- src/main/java/com/example/courier304project/repository/PickupRepository.java
@@ -1,11 +1,0 @@
-package com.example.courier304project.repository;
-
-import com.example.courier304project.entity.Pickup;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PickupRepository extends JpaRepository<Pickup,Long> {
-
-
-}
